Move show running app on the top setting to bottomsheet
