+++ feature/settings/src/main/res/values-zh-rTW/strings.xml
@@ -53,6 +53,5 @@
     <string name="options_on">啟用</string>
     <string name="options_off">關閉</string>
     <string name="report_bugs_or_submit_ideas">問題反饋 &amp; 我有新點子</string>
-    <string name="show_running_apps_on_top">置頂運行中應用</string>
     <string name="designers_homepage">設計師主頁</string>
 </resources>
