+++ feature/settings/src/main/res/values/strings.xml
@@ -61,6 +61,5 @@
     <string name="options_on">On</string>
     <string name="options_off">Off</string>
     <string name="report_bugs_or_submit_ideas">Report bugs or submit ideas</string>
-    <string name="show_running_apps_on_top">Show running apps on top</string>
     <string name="designers_homepage">Designer\'s homepage</string>
 </resources>
