+++ feature/sort/src/main/res/values-zh-rCN/strings.xml
@@ -28,4 +28,7 @@
     <string name="name">姓名</string>
     <string name="install_date">安装日期</string>
     <string name="last_updated">最后更新</string>
+    <string name="show_running_apps_on_top">置顶运行中应用</string>
+    <string name="on">开启</string>
+    <string name="off">关闭</string>
 </resources>
