+++ feature/sort/src/main/res/values-zh-rTW/strings.xml
@@ -28,4 +28,7 @@
     <string name="name">姓名</string>
     <string name="install_date">安裝日期</string>
     <string name="last_updated">最後更新</string>
+    <string name="show_running_apps_on_top">置頂運行中應用</string>
+    <string name="on">開啟</string>
+    <string name="off">關閉</string>
 </resources>
